Add the Developer Guide (#1)
* Add the Developer Guide
This PR adds the developer guide to the repository. This is only a draft of
the guide, and parts of it are not yet fully reviewed. However, I feel that
it's time to start being able to review the guide in its Markdown form rather
than its HTML format, since there are differences in presentation.

Co-authored-by: Nate Prewitt <nate.prewitt@gmail.com>